fix warnings

+++ packages/protocol/solidity-contracts/Distribution/DistributionFace.sol
@@ -23,9 +23,9 @@ pragma experimental "v0.5.0";
 /// @author Gabriele Rigo - <gab@rigoblock.com>
 interface DistributionFace {
 
-    event Subscription(address indexed buyer, address indexed distributor, uint amount);
+    event Subscription(address indexed buyer, address indexed distributor, uint256 amount);
 
     function subscribe(address _pool, address _distributor, address _buyer) external payable;
-    function setFee(uint _fee, address _distributor) external;
-    function getFee(address _distributor) external view returns (uint);
+    function setFee(uint256 _fee, address _distributor) external;
+    function getFee(address _distributor) external view returns (uint256);
 }
