+++ packages/protocol/solidity-contracts/Drago/DragoFace.sol
@@ -28,15 +28,15 @@ interface DragoFace {
     function() external payable;
     function buyDrago() external payable returns (bool success);
     function buyDragoOnBehalf(address _hodler) external payable returns (bool success);
-    function sellDrago(uint _amount) external returns (bool success);
-    function setPrices(uint _newSellPrice, uint _newBuyPrice)  external;
+    function sellDrago(uint256 _amount) external returns (bool success);
+    function setPrices(uint256 _newSellPrice, uint256 _newBuyPrice)  external;
     function changeMinPeriod(uint32 _minPeriod) external;
-    function changeRatio(uint _ratio) external;
-    function setTransactionFee(uint _transactionFee) external;
+    function changeRatio(uint256 _ratio) external;
+    function setTransactionFee(uint256 _transactionFee) external;
     function changeFeeCollector(address _feeCollector) external;
     function changeDragoDao(address _dragoDao) external;
-    function depositToExchange(address _exchange, uint _amount) external;
-    function withdrawFromExchange(address _exchange, uint _amount) external;
+    function depositToExchange(address _exchange, uint256 _amount) external;
+    function withdrawFromExchange(address _exchange, uint256 _amount) external;
     function setInfiniteAllowance(address _tokenTransferProxy, address _token) external;
     function SetMultipleAllowances(address _tokenTransferProxy, address[] _token) external;
     function operateOnExchange(address _exchange, bytes _assembledTransaction) external;
@@ -46,11 +46,11 @@ interface DragoFace {
 
     // PUBLIC CONSTANT FUNCTIONS
 
-    function balanceOf(address _who) external view returns (uint);
+    function balanceOf(address _who) external view returns (uint256);
     function getEventful() external view returns (address);
-    function getData() external view returns (string name, string symbol, uint sellPrice, uint buyPrice);
-    function calcSharePrice() external view returns (uint);
-    function getAdminData() external view returns (address, address feeCollector, address dragoDao, uint ratio, uint transactionFee, uint32 minPeriod);
+    function getData() external view returns (string name, string symbol, uint256 sellPrice, uint256 buyPrice);
+    function calcSharePrice() external view returns (uint256);
+    function getAdminData() external view returns (address, address feeCollector, address dragoDao, uint256 ratio, uint256 transactionFee, uint32 minPeriod);
     function totalSupply() external view returns (uint256);
     function getKycProvider() external view returns (address);
     function getVersion() external view returns (string);
